requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.0
requests==2.31.0
werkzeug==2.3.7
PyPDF2==3.0.1
python-docx==0.8.11
Pillow>=10.0.0
python-multipart==0.0.6
pytesseract==0.3.10
gunicorn==21.2.0
flask-socketio==5.6.0
python-socketio==5.12.0
eventlet==0.33.3
pyspellchecker==0.8.4
